Remove the Facebook login feature for now. (see #628)
The login-with-Facebook feature has not been working in Production
for some time, which we just discovered recently. Because of this,
and because of the inability to install it from a local fork of
the project we were using for this feature, it is simply removed
from the site for now, with a possible return upon future review.